requirements.txt
alembic==1.14.0
blinker==1.9.0
click==8.1.8
faiss-cpu==1.9.0.post1
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
Mako==1.3.8
MarkupSafe==3.0.2
numpy==2.2.1
packaging==24.2
PyMySQL==1.1.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
